Fix clippy warnings
  - Use is_ok() instead of if let Ok(_)
  - Remove unnecessary casts
  - Use push('\n') instead of push_str

+++ crates/obsfs-plugins/src/procsys/fs.rs
@@ -106,10 +106,10 @@ fn get_fs_stats(mount_point: &str) -> anyhow::Result<FsStats> {
 
     let stat = unsafe { stat.assume_init() };
 
-    let block_size = stat.f_frsize as u64;
-    let total = stat.f_blocks as u64 * block_size;
-    let available = stat.f_bavail as u64 * block_size;
-    let free = stat.f_bfree as u64 * block_size;
+    let block_size = stat.f_frsize;
+    let total = stat.f_blocks * block_size;
+    let available = stat.f_bavail * block_size;
+    let free = stat.f_bfree * block_size;
     let used = total - free;
 
     Ok(FsStats {
